+++ CJBShowItemSellPrice/ModEntry.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using CJBShowItemSellPrice.Framework;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework.Input;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
 using StardewModdingAPI.Utilities;
@@ -202,15 +201,15 @@ namespace CJBShowItemSellPrice
             Vector2 outerSize = innerSize + new Vector2((borderSize + padding) * 2);
 
             // get position
-            float x = (Mouse.GetState().X / Game1.options.zoomLevel) - offsetFromCursor.X - outerSize.X;
-            float y = (Mouse.GetState().Y / Game1.options.zoomLevel) + offsetFromCursor.Y + borderSize;
+            float x = Game1.getMouseX() - offsetFromCursor.X - outerSize.X;
+            float y = Game1.getMouseY() + offsetFromCursor.Y + borderSize;
 
             // adjust position to fit on screen
             Rectangle area = new Rectangle((int)x, (int)y, (int)outerSize.X, (int)outerSize.Y);
-            if (area.Right > Game1.viewport.Width)
-                x = Game1.viewport.Width - area.Width;
-            if (area.Bottom > Game1.viewport.Height)
-                y = Game1.viewport.Height - area.Height;
+            if (area.Right > Game1.uiViewport.Width)
+                x = Game1.uiViewport.Width - area.Width;
+            if (area.Bottom > Game1.uiViewport.Height)
+                y = Game1.uiViewport.Height - area.Height;
 
             // draw tooltip box
             IClickableMenu.drawTextureBox(spriteBatch, Game1.menuTexture, this.TooltipSourceRect, (int)x, (int)y, (int)outerSize.X, (int)outerSize.Y, Color.White);