+++ CJBCheatsMenu/Framework/CheatsMenu.cs
@@ -307,8 +307,8 @@ namespace CJBCheatsMenu.Framework
             // set dimensions
             this.width = (this.IsAndroid ? 750 : 800) + IClickableMenu.borderWidth * 2;
             this.height = (this.IsAndroid ? 550 : 600) + IClickableMenu.borderWidth * 2;
-            this.xPositionOnScreen = Game1.viewport.Width / 2 - (this.width - (int)(Game1.tileSize * 2.4f)) / 2;
-            this.yPositionOnScreen = Game1.viewport.Height / 2 - this.height / 2;
+            this.xPositionOnScreen = Game1.uiViewport.Width / 2 - (this.width - (int)(Game1.tileSize * 2.4f)) / 2;
+            this.yPositionOnScreen = Game1.uiViewport.Height / 2 - this.height / 2;
 
             // show close button on Android
             if (this.IsAndroid)