+++ CJBCheatsMenu/Framework/Components/CheatsOptionsNPCSlider.cs
@@ -85,9 +85,10 @@ namespace CJBCheatsMenu.Framework.Components
         /// <param name="spriteBatch">The sprite batch being drawn.</param>
         /// <param name="slotX">The X position at which to draw, relative to the bounds.</param>
         /// <param name="slotY">The Y position at which to draw, relative to the bounds.</param>
-        public override void draw(SpriteBatch spriteBatch, int slotX, int slotY)
+        /// <param name="context">The menu drawing the component.</param>
+        public override void draw(SpriteBatch spriteBatch, int slotX, int slotY, IClickableMenu context = null)
         {
-            base.draw(spriteBatch, slotX + this.GetOffsetX(), slotY);
+            base.draw(spriteBatch, slotX + this.GetOffsetX(), slotY, context);
 
             Color tint = this.greyedOut ? (Color.White * 0.5f) : Color.White;